bindings/kotlin: added general way for serialization
